Reduced CPU usage of memory table.
`tui::Table` can't reuse a formatted table across multiple draws, and
each `tui::Cell` performs several allocations.  This has really bad
performance for large tables.  It looks like this is by design in
`tui`.  Even though the text itself isn't copied, the preprocessing
done when generating a `tui::Cell`, iterating through each line and
allocating a vector of spans, doesn't have an obvious way to be
avoided.

There's also no way to set `TableState::offset`, so the table couldn't
be generated as a smaller size, with the window moving alongside it.
Doing so would require setting the `offset` to mimic a full-sized
table.

Instead, making a table row for each location, but only filling the
cells of the table if within range of the selected line.