Added title for memory table, including entry point
First step in showing exploration of more than just the `[stack]`
space.  Title includes name of memory region, entry point, and offset
relative to the entry point.  The entry point for `[stack]` is the
largest address (first used) in the stack.  The entry point for other
regions will be the pointed-to location that was followed into that
region.