Performance improvements for memory table
The `Table` widget in `ratatui` requires a vector of rows.  Even if
rows are empty whenever they are outside of the visible area, the
allocation of each row can still be a bottleneck.

By replacing the `Table` widget with a `DynamicTable` widget, a cell
is only generated when it is going to be displayed.  This has some
limitations, as the height of each row must be determined prior to its
contents, but this works well for the `MemoryTable`.

This change provides a ~6x improvement in performance (20 ms per
frame, down from 120 ms).