Cleanup, handle all index offsets in conversion to `usize`
Some CLI indices start at zero (`#Strings` and `#Blob`), while others
start at one (`#GUID` and table indices) in order to use zero for an
absent value.  This offset is now handled in the conversion to Rust
`usize`, rather than at the usage of the index, and uses `Option` to
represent absent values.