Start collecting MetadataLayout struct
There's some lifetime issues arising when the `&'b Metadata<'a>`
object escapes the closure used for an iterator.  This should be
resolvable by having a known storage location for the layout
information, separate from the references held by `Metadata<'a>`.  As
it is, `Metadata<'a>` holds both owned objects and references.