Performance improvement, module search
Searching for the runtime module was taking too long on startup.
The main bottleneck was from hashing potential pointer values to see
if they were a pointer to an IL method definition.  Using the range of
possible pointer values to filter out values before hashing gave a ~3x
performance improvement (4.5 seconds per search down to 1.5).

Probably should move the `PersistentState` class to be owned by the
`TuiExplorer`, so that the search only needs to be performed once.
That would save another 3 seconds of startup time.