Return Pointer from FieldDescription::location
Previously, it returned `Range<Pointer>`, which had the incorrect size
when the field is a struct.  Using only the information in the
FieldDescription, retrieving the size of a field within the
base class was possible when the field is a class/primitive type, but
not when the field is a struct.

The size of the field should be determined using
`RuntimeType::size_bytes`.