Bugfix, do not requrire generic types of the base class in Signature
Prior to this commit, both the base and parent's classes were compared
against the metadata signature.  As a result, instantiations of a
`class Child<T>: Base<T>` would never be found, since the signature
has one type argument, and the method table was erroneously required
to have two type arguments.