Bugfix, avoid printing of base class's generics
Same root cause as the previous commit, but appearing when printing
a TypeHandle.  I probably should just remove the `generic_types`
function altogether.  It was originally there to handle generic fields
where the generic argument was in a base class, but that is now better
handled by providing a field's parent in `iter_instance_fields()`.