Rudimentary auto-fishing
Uses `enigo` to send keystrokes and `active_win_pos_rs` to see if
Stardew Valley is the active window.  However, as far as I can tell,
neither has the ability to send a keystroke to a specific window
handle, or to check if a specific window handle is currently active.

So, I may need to drop down a level of abstraction and interact with
the X11 libs directly.