Handle missing method table for ValueType
This can occur when attempting to find the method table for an
instantiated generic type, but no instances of the type have been
constructed.  For example, if `MyGeneric<i32>` only appears inside
`Array<MyGeneric<i32>>`, and all such arrays are empty, then the
method table for `MyGeneric<i32>` may yet to be generated in the
remote process.