Make zero-width space opt-in when printing expression
The zero-width space characters are inserted to indicate good
locations for line breaks.  However, it's not necessary to include
them in all cases, and it makes it a bit tedious to include their
exact location in test cases.

This commit updates the printing to only include the zero-width spaces
when explicitly requested.