Bugfix, handle EEClass located near end of allocation
The `EEClass` structure has a common prefix, a middle section that
varies by type, and a common suffix.  The information I'm
extracting is in the common prefix and suffix, but the location of the
suffix depends on the size of the intermediate section.

Previously, I had just read out an extra 256 bytes, so that I'd get
the entire `EEClass` regardless of which subtype it is.  However, if
the `EEClass` is located near the end of a memmap region, this may be
an invalid read.  Therefore, updating to instead read 256 bytes, or up
to the end of the containing memmap region, whichever is shorter.