Instruction::NativeFunctionCall with arg list, optional output
Previously, the registers holding arguments for a
`Instruction::NativeFunctionCall` needed to be located in sequential
order, and could not use `VMArg::Const` arguments.  This commit
updates the instruction to instead hold a vector of `VMArg` instances,
removing both restrictions.

In addition, the output argument is now optional for
`Instruction::NativeFunctionCall`.  Because this instruction may have
side-effects, it doesn't strictly require an output location to be
useful.