Cache reads from VM to improve startup time
The majority of the bot's startup time is in reading values from
Stardew Valley.  Some of this may be avoided, since there's some extra
system calls (e.g. `ProcessVirtualMemoryIO::ensure_process_exists`
takes a decent percentage of the time waiting for POSIX `kill` to
return), but most of it is just because I'm doing too many small
reads.

Whenever reading values from Stardew's RAM, the caching now reads ~2
kB around the value being read out, caching up to 100 such chunks to
use for later reads.  This caching is per function evaluation, so the
next frame will still get fresh values.  This improved the time
required for the initial read from ~50 seconds to ~10 seconds.

This may be improved in the future through some static analysis, to
batch reads that are known to be upcoming.