Skip first and last waypoint when pathfinding
Omit the first and last waypoints.  These are the tiles that include
the player's current position and the goal position.  Rather than
going to these waypoints, it's better to go directly to the next
position, which avoids accidental backtracking.

Previously, while the last waypoint was omitted, the first waypoint
was included, even if this required moving away from the goal
location.