Perform all reads through MemoryReader::read_exact
Previously, the `Pointer::read_exact` method could read bytes from a
remote process.  However, this required re-initializing the
`ProcessVirtualMemoryIO` instance for each read.  When initialized,
the `ProcessVirtualMemoryIO` checks that the PID currently exists.
This caused an extra call to `kill(pid, 0)` for each memory read.

Avoiding this re-initialization of `ProcessVirtualMemoryIO` improved
the runtime by ~20%.  Not enough to compensate for the slowdown in
reading location information, but a step in the right direction.