Use binary search to find containing region
When determining how large of a read can safely be performed on the
remote process, the containing region is used as an outer bound.
Iterating over the known regions to find the containing region was
slow enough that it showed up in the profile.

Replacing the linear search with a binary search gives another 5-10%
performance increase.