Implement Downcast in terms of IsSubclassOf
Previously, the read of the method table pointer couldn't be batched
by `MergeParallelReads`, since it was internal to the implementation
of `ExprKind::PhysicalDowncast`.  This exposes the subclass check as
`ExprKind::IsSubclassOf`, such that the read of the method table
pointer can be batched.

Currently, this change makes the compilation take 2x as long, with no
significant change to the runtime performance.  Therefore, the
`ExprKind::PhysicalDowncast` is left in for now, and can be re-enabled
with the environment variable `USE_PHYSICAL_DOWNCAST` for
investigation.