Implement GoToActionGoal, for navigating to a special-action tile
For example, "Carpenter" for navigating to Robin's shop.

Does not currently consider whether the location is accessible or
usable.  A location may be inaccessible if it requires going through a
`LockedDoorWarp` outside of the unlocked hours.  Each special-action
may have additional actions required for it to be usable, such as
the "Blacksmith" action requiring Clint to be at the desk.