For room-to-room movement, track index of the Location during search
Previously, the node was specified as a tuple of the room name and the
location within the room.  Now, the room name is replaced by the index
of the room within `GameState.locations`.  This avoids string copies,
and prepares for initializing the `LocalMovementGoal` with more
information about the door to be used.