First day optimizations, scythe path to water, leave clay on ground
Since there's some extra time between finishing the first round of
clay farming and the Saloon getting unlocked, it can be spent clearing
a path to the nearest watering hole.  This avoids the circuitous route
that would otherwise be used later on.

The bot then arrives a bit late to the Saloon, which can be avoided by
skipping the cleanup step on the first round of clay farming.  Since
we're coming back to the beach anyways, there's no point in picking up
every stray piece of clay after the first round.