Pathfinding bugfix for A* search
There are two places where the `dist_plus_heuristic` field is checked
to see if it is the best path to a tile.  First, when pushing into the
`to_visit` btree, to only collect the best known route.  Second, when
popping from the `to_visit` btree, in case the route was superseded
by a better route.

Previously, the condition for pushing into `to_visit` was correct,
but the condition when popping an entry out was reversed.  This is
what caused the small diagonal movements just prior to rounding a
corner, since the diagonal movement would be pushed into `to_visit`,
and would then overwrite the best path.