Increase the radius for reading objects around the player
For larger structures, such as resource clumps, this avoids having
them flicker out of existence.  I think this is caused by the resource
clump being pruned when the center is within 5 tiles of the player,
but only being read out when the top-left corner is within 6 tiles of
the player.  For some edge cases, this can result in the ResourceClump
being pruned, but not being restored.

Probably should update the re-read logic for ResourceClumps to use the
same threshold in both `read_location_resource_clumps` and
`apply_delta`.