Update UseItemOnTile to reduce time required for scythe
Previously, animation canceling was only applied for items that are
defined as tool (start with `(T)`), since those are easiest to
determine when the tool has had an effect.  The scythe is categorized
as a weapon, not a tool, and has an effect on each frame of its
animation.

This commit updates the animation canceling in `UseItemOnTool` to
cancel out of weapon animations, if the animation has already reached
the last frame on which it deals damage.