Allow bot to cross the repaired beach bridge
There are two types of water tiles on the map.  Those marked with
"Water", and those marked with both "Water" and "Passable".  Since
this is on the "Back" layer of the map, the "Passable" flag means that
a tile is impassable.  Water that can be walked on (just the "Water"
flag) is used for the majority of tiles, while water that cannot be
walked on (both the "Water" and "Passable" flags) is only used for the
border of water areas.

Previously, all water tiles were marked as blocked, regardless of
whether they had the "Passable" flag.  This had the correct behavior
for pathfinding, because the walkable water couldn't be accessed
anyways.

Turns out, there is one instance in which a walkable water tile is
accessible: the repaired bridge at the beach.  In order to allow the
bot to walk across the repaired bridge, the pathfinding must not treat
water tiles as being blocked, as this would prevent walking across the
bridge.