Refactor LocalMovementGoal to use a TileMap<Direction>
Previously, the `LocalMovementGoal` performed an A* search from the
player's current location to the target location, then tracked the
path as a series of waypoints to be followed.  This caused unnecessary
backtracking if the movement was interrupted, and the player was no
longer near the waypoint when the movement was resumed.