Tweak movement to prefer being in center of tile
When using the waypoint-based movement, the bot naturally aligned the
player to the center of each tile, since that was the location of each
waypoint.  When using the direction map, the bot tends to instead
align the player along the edge of a tile, since that is where the
preferred direction changes.  This results in the bot bumping into
corners of obstacles, since the collision box edges beyond the edge of
the player's tile.

This commit updates the movement to move the player toward the center
of a tile, by adding diagonal movement whenever the
`Pathfinding.direction_map` indicates motion in a cardinal direction,
and the player is on a tile's edge.  This is a bit of a hack, and may
be replaced by a more complete solution that divides each tile up into
a 3x3 grid for the purpose of pathfinding.