Bugfix for thresholds in MineNearbyOre interrupt
Before the `MineNearbyOre` interrupt was allowed to break rocks on the
way to valuable ore, the distance to any ore included the final step
onto the ore-containing tile, but did not include the cost of mining
the ore itself.  When the `MineNearbyOre` interrupt was updated to
allow breaking rocks that were in the way, this also caused the
distance to the ore-containing tiles to include the cost of minding
the ore itself.

As a result, the cutoff of `dist < 4` tiles with a cost of 2 tiles could
only be met if the player was directly adjacent to the tile with the
ore.

This commit updates the thresholds to have an offset, canceling out
cost associated with mining the ore itself.