Better prioritization for PlantCropsGoal
Previously, use of the watering can would be delayed if opportunistic
clay farming was enabled.  This ensured that, if the clay farming was
necessary to get cash before planting seeds, the slower step of going
to/from a source of water would be skipped.  However, this means that
the player needs to return to the tiles later to water them, which can
slow down the process overall.

Now, check both the opportunistic clay farming and that a stop time
has been defined to determine whether watering should be delayed.