Improve the re-plan interrupt for MineSingleLevel
Previously, would only replan if the number of objects in the level
increased, which was usually caused by an enemy dropping a ladder.
However, this could still leave the bot stuck in some cases, since
`MineNearbyOre` could have decreased the number of objects prior to
the ladder dropping.

With this commit, also re-plan if half a second has passed.