Correct argument order in AsGridPos implementations
The argument order is `(width,height)`, but some locations instead
passed `(height,width)`.  This seems like it should have broken
quite a lot of pathfinding checks, so I'm not quite sure how it was
working previously.  Possibly that so long as the same index type was
used consistently for a given `TileMap`, and since the tiles are
almost exclusively represented with `Vector<isize>`, the values would
be stored in a transposed location, but it would be the same
transposed position for both read/write.