Avoid edge case in register assignment
Previously, a reduction's accumulator was owned by the parent scope,
requiring a reduction was required to preserve the output of the
previous loop.  In some cases, such as different usage of the
accumulator along if/else branches, this prevented generation of a
valid register assignment.

Since each iteration of the reduction loop must produce the
accumulator value for the next iteration, the accumulator may be
considered to be owned by the reduction function.  This allows it to
be moved into the output location, avoiding these edge cases.