Check more than just the top CancelIf
I think this is the cause of the occasional misfire of the
infinite-loop detection.  If there's more than one cancellation that
could be triggerred, then the first loop would only trigger the
top-most cancellation.  A goal between the two interrupts could then
produce subgoals.  When the second interrupt triggers, it would then
look like the subgoals had all been completed, but the triggering goal
is no longer present.