Bugfix for skipping long waits during FishOnceGoal
When fishing with the bamboo rod, the bot skips waits that would be
too long and re-casts.  This behavior was broken with some of the
automatic button release logic.  When the conditional that
would skip the wait is met, the `GameAction::ReleaseTool` would have
no effect if the use-tool button was not already being held down.

Fixed by instead sending a left-click for early-skipping.