Faster ladder-searching in case where no ladder-rocks exist
The number of prior stones that must be mined prior to a stone being
able to produce a ladder is determined by the RNG roll on that stone.
If no stones would currently produce a ladder, as stones are mined
away, the stones with lower RNG roll will be the first ones that will
be capable of producing a ladder.  Therefore, prefer to preserve these
low-RNG-roll stones until they will produce a ladder.