Temporarily disable the reading of player buffs
Reading the buffs currently has the same problem that occurs for
reading of floating items, but is harder to work around.  Until an
instance is constructed, there is no method table for the generic type
``"System.Collections.Generic.Dictionary`2"<System.String,
StardewValley.Buff>``.  An instance is constructed when the player
first gets a buff from food/drink, but that isn't always convenient to
trigger.

Fixing this will require changing the lowering model from
`SymbolicGraph`/`SymbolicExpr` to `VirtualMachine`/`Instruction`, such
that the method table lookup occurs on attempting to access a field,
rather than when planning/compiling the expression.