Use tie-breaker for equidistant rocks to mine
Since the rocks that may be worth mining are collected in a `HashMap`,
two rocks that have the same distance from the player will cause an
arbitrary element to be returned.  Since `MineSingleLevel` will
periodically re-plan to avoid getting stuck, this can cause the bot to
get stuck, unable to reach either target rock before re-planning
causes it to go to the other one.

Using the tile's location as a tie-breaker avoids this edge case.