Improve stutter-stepping in StepCountForLuck goal
Previously, would stop movement whenever the player is moving, and the
step count has reached the target value.  Now, it waits until the
animation frame prior to incrementing of the `stepsTaken` counter.
This allows the bot to move for about 250 ms, followed by 16 ms of
standing still, significantly increasing the average speed during this
time.