Provide vector name in error message when attempting to push None
This has probably been one of the more tedious error messages to run
into, because it sometimes just needed a filter prior to collecting,
but that could be the default because it also could be an actual error
in unpacking.  But the error didn't occur until runtime, and by that
point the name of the collection was no longer available.