When reserving items for bundles, check if non-base quality required
Previously, reservation of items for bundles only checked for exact
matches of the `ItemId`.  If no base-quality item was available, and a
bundle would instead be completed using a higher-quality item, that
higher-quality item would not be correctly reserved, and could be
erroneously eaten for stamina.

This commit updates the `InventoryGoal.stamina_recovery_slots`
handling to check for reserved items of a higher quality.