Improve type inference for iter.reduce(intiial, reduction)
Previously, the reduction would always be inferred to be the same type
as `initial`, but the initial value may not have a known type.
Now, it also checks the first parameter of the reduction function and
the output of the reduction function, which may also be used to infer
the result of `iter.reduce(intiial, reduction)`.