Prefer smelting whichever type of bar is least available
Previously, this was implicit from the choice of which region to mine
in.  However, if there is insufficient coal to smelt all collected
ore, there may be extra farming of the iron levels for dust sprites,
resulting in an excess of iron ore.  In that case, we still want to
smelt copper bars to keep an equal number of copper and iron bars for
sprinklers.