Use linear search for bindings in interpreter
Provides a ~20% improvement over the `HashMap` in benchmmarks, though
this may be biased from the small number of variables bindings present
in the benchmarks as compared to the bot itself.  Either way, the
interpreter is still significantly slower than the bytecode, so the
functionality improvements may need to be made on the bytecode itself,
rather than the interpreter.