Persist ExprKind::LazyStatic results across VM executions
Initially, the `ExprKind::LazyStatic` definitions were only saved for
later use in the same VM execution.  This avoided the repeated lookup
of struct layouts within a single frame, but still required checking
the layouts on subsequent frames.

With this commit, `ExprKind::LazyStatic` saves values to the
`VirtualMachine`, restoring them for use on the next execution.